Create BIG_controle.py
Python script for connecting an AFAS-environment of a customer to the BIG-register database of healthcare professionals.
The script retrieves the information of the employees scheduled for checkups. This information is passed to the webservice of the BIG-register and the response is handled, formatted and send back to the employee dossier in the AFAS-environment.