Create FileConversion_CreateTransactionDatabase.py
One of my projects was a file conversion. Two AFAS environments were merged into one. All the files of one environment had to to be retrieved through the AFAS API and be send to the second environment. The challenge was to send each file to the correct 'location' in the second environment, i.e. to the correct employee, dossier number and other properties. 
Tabular data from both the old and new environment was combined to achieve a mapping for each file. Since some tables were too large to be handled in Excel (up to 300.000 lines and 30 columns), both Pandas and SQLite was used for processing the data and ultimately creating a single transaction table.
A separate script was used to read the transaction table for the downloading and uploading of the files themselves.